+++ src/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
-import { auth } from './server/firebase';
 import { Navigate } from 'react-router-dom';
+import { auth } from './server/firebase';
 
 export default function ProtectedRoute({
   children,
