+++ src/components/layout/Tweet.tsx
@@ -1,9 +1,9 @@
+import { User } from 'firebase/auth';
+import { DotsHorizontalIcon, PersonIcon } from '@radix-ui/react-icons';
 import { AspectRatio } from '@/components/ui/aspect-ratio';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { Card, CardContent } from '@/components/ui/card';
-import { ITweet } from '@/hooks/useLoadTweet';
-import { DotsHorizontalIcon, PersonIcon } from '@radix-ui/react-icons';
-import { User } from 'firebase/auth';
+import { ITweet } from '@/hooks';
 
 interface Props extends ITweet {
   user?: User | null;
